+++ src/query/service/src/interpreters/interpreter_copy_into_location.rs
@@ -29,8 +29,8 @@ use log::debug;
 use crate::interpreters::common::check_deduplicate_label;
 use crate::interpreters::Interpreter;
 use crate::interpreters::SelectInterpreter;
-use crate::pipelines::builders::build_append2table_with_commit_pipeline;
 use crate::pipelines::PipelineBuildResult;
+use crate::pipelines::PipelineBuilder;
 use crate::sessions::QueryContext;
 use crate::sessions::TableContext;
 use crate::sql::plans::CopyIntoLocationPlan;
@@ -111,7 +111,7 @@ impl CopyIntoLocationInterpreter {
             default_values: None,
         };
         let to_table = StageTable::try_create(stage_table_info)?;
-        build_append2table_with_commit_pipeline(
+        PipelineBuilder::build_append2table_with_commit_pipeline(
             self.ctx.clone(),
             &mut build_res.main_pipeline,
             to_table,
